Extract out a test generator so it can be used in other places
